Crate pinnacle_api

The Rust implementation of Pinnacle’s configuration API.

This library allows you to interface with the Pinnacle compositor and configure various aspects like input and the tag system.

Configuration

1. Create a cargo project

To create your own Rust config, create a Cargo project in ~/.config/pinnacle.

2. Create metaconfig.toml

Then, create a file named metaconfig.toml. This is the file Pinnacle will use to determine what to run, kill and reload-config keybinds, an optional socket directory, and any environment variables to give the config client.

In metaconfig.toml, put the following:

# `command` will tell Pinnacle to run `cargo run` in your config directory.
# You can add stuff like "--release" here if you want to.
command = ["cargo", "run"]

# You must define a keybind to reload your config if it crashes, otherwise you'll get stuck if
# the Lua config doesn't kick in properly.
reload_keybind = { modifiers = ["Ctrl", "Alt"], key = "r" }

# Similarly, you must define a keybind to kill Pinnacle.
kill_keybind = { modifiers = ["Ctrl", "Alt", "Shift"], key = "escape" }

# You can specify an optional socket directory if you need to place the socket Pinnacle will
# use for configuration in a different place.
# socket_dir = "your/dir/here"

# If you need to set any environment variables for the config process, you can do so here if
# you don't want to do it in the config itself.
[envs]
# key = "value"

3. Set up dependencies

In your Cargo.toml, add pinnacle-api as a dependency:

# Cargo.toml

[dependencies]
pinnacle-api = { git = "https://github.com/pinnacle-comp/pinnacle" }

4. Set up the main function

In main.rs, change fn main() to async fn main() and annotate it with the pinnacle_api::config macro:

#[pinnacle_api::config]
async fn main() {}

5. Begin crafting your config!

You can create the API modules with ApiModules::new:

use pinnacle_api::ApiModules;

let ApiModules {
    ..
} = ApiModules::new();

Most modules are copy-able unit structs, so you can also use them directly:

let _ = pinnacle_api::window::Window.get_all();
pinnacle_api::pinnacle::Pinnacle.quit();

You can peruse the documentation for things to configure.

Re-exports

pub use snowcap_api;

pub use tokio;

Modules

input

Input management.

layout

Layout management.

output

Output management.

pinnacle

Compositor management.

process

Process management.

render

Rendering management.

signal

Compositor signals.

snowcap

The Snowcap widget system. // TODO: these docs

tag

Tag management.

util

Utility types.

window

Window management.

Macros

batch_boxed

A convenience macro to batch API calls in different concrete futures.

batch_boxed_async

The async version of batch_boxed.

Structs

ApiModules

A struct containing all of the config module structs.

Functions

connect

Connects to Pinnacle and builds the configuration structs.

listen

Listen to Pinnacle for incoming messages.

set_default_tracing_subscriber

Sets the default tracing_subscriber to output logs.

Attribute Macros

config

Transform the annotated function into one used to configure the Pinnacle compositor.