pinnacle_api::input

Function set_repeat_rate

pub fn set_repeat_rate(rate: i32, delay: i32)

Sets the keyboard’s repeat rate.

This allows you to set the time between holding down a key and it repeating as well as the time between each repeat.

Units are in milliseconds.

Examples

// Set keyboard to repeat after holding down for half a second,
// and repeat once every 25ms (40 times a second)
input::set_repeat_rate(25, 500);