pinnacle_api::input::libinput

Struct DeviceHandle

pub struct DeviceHandle { /* private fields */ }

A libinput device.

Implementations

impl DeviceHandle

pub fn capabilities(&self) -> Capability

Gets the capabilities of this device.

pub async fn capabilities_async(&self) -> Capability

Async impl for Self::capabilities.

pub fn name(&self) -> String

Gets this device’s name.

pub async fn name_async(&self) -> String

Async impl for Self::name.

pub fn product_id(&self) -> u32

Gets this device’s product id.

pub async fn product_id_async(&self) -> u32

Async impl for Self::product_id.

pub fn vendor_id(&self) -> u32

Gets this device’s vendor id.

pub async fn vendor_id_async(&self) -> u32

Async impl for Self::vendor_id.

pub fn device_type(&self) -> DeviceType

Gets this device’s DeviceType.

pub async fn device_type_async(&self) -> DeviceType

Async impl for Self::device_type.

pub fn map_to_output(&self, output: &OutputHandle)

Maps the absolute input from this device to the corresponding output.

This will cause touch input from this device to map proportionally to the area of an output. For example, tapping in the middle of the device will generate a tap event at the middle of the output.

This only affects devices with touch capability.

If you want to map the device to an arbitrary region, see Self::map_to_region.

pub fn map_to_region(&self, region: Rect)

Maps the absolute input from this device to the corresponding region in the global space.

This will cause touch input from this device to map proportionally to the given region within the global space. For example, tapping in the middle of the device will generate a tap event at the middle of the region. This can be used to map a touch device to more than one output, for example.

This only affects devices with touch capability.

If you want to map the device to a single output, see Self::map_to_output.

pub fn set_accel_profile(&self, accel_profile: AccelProfile)

Sets this device’s acceleration profile.

pub fn set_accel_speed(&self, accel_speed: f64)

Sets this device’s acceleration speed.

pub fn set_calibration_matrix(&self, calibration_matrix: [f32; 6])

Sets this device’s calibration matrix.

pub fn set_click_method(&self, click_method: ClickMethod)

Sets this device’s click method.

pub fn set_disable_while_typing(&self, disable_while_typing: bool)

Sets whether or not this device is disabled while typing.

pub fn set_left_handed(&self, left_handed: bool)

Sets this device to left-handed or not.

pub fn set_middle_emulation(&self, middle_emulation: bool)

Sets whether or not middle emulation is enabled.

pub fn set_rotation_angle(&self, rotation_angle: u32)

Sets this device’s rotation angle.

pub fn set_scroll_button(&self, scroll_button: u32)

Sets this device’s scroll button.

pub fn set_scroll_button_lock(&self, scroll_button_lock: bool)

Sets whether or not the scroll button locks on this device.

pub fn set_scroll_method(&self, scroll_method: ScrollMethod)

Sets this device’s scroll method.

pub fn set_natural_scroll(&self, natural_scroll: bool)

Enables or disables natural scroll on this device.

pub fn set_tap_button_map(&self, tap_button_map: TapButtonMap)

Sets this device’s tap button map.

pub fn set_tap_drag(&self, tap_drag: bool)

Enables or disables tap dragging on this device.

pub fn set_tap_drag_lock(&self, tap_drag_lock: bool)

Sets whether or not tap dragging locks on this device.

pub fn set_tap(&self, tap: bool)

Enables or disables tap-to-click on this device.

pub fn set_send_events_mode(&self, send_events_mode: SendEventsMode)

Sets this device’s send events mode.