pinnacle_api

Module layout

Layout management.

Read the wiki page for more information.

Modules

generators

Various builtin generators.

Structs

Gaps

Gaps around a layout node.

LayoutArgs

Arguments from an incoming layout request.

LayoutNode

A single node of a layout tree.

LayoutRequester

A struct that can request layouts.

LayoutResponse

A response to a layout request containing a layout tree.

Enums

LayoutDir

A layout direction.

Traits

LayoutGenerator

Types that can generate layouts by computing a tree of LayoutNodes.

Functions

manage

Manages layout requests from the compositor.