pinnacle_api::layout
Module generators

Various builtin generators.

Structs

Corner

A LayoutGenerator that has one main corner window and a horizontal and vertical stack flanking it on the other two sides.

Cycle

A LayoutGenerator that keeps track of layouts per tag and provides methods to cycle between them.

Dwindle

A LayoutGenerator that lays out windows in a shrinking fashion towards the bottom right corner.

Fair

A LayoutGenerator that attempts to layout windows such that they are the same size.

Line

A LayoutGenerator that lays out windows in a line.

MasterStack

A LayoutGenerator that has one master area to one side and a stack of windows next to it.

Spiral

A LayoutGenerator that lays out windows in a spiral.

Enums

CornerLocation

Which corner the corner window will in.

MasterSide

Which side the master area will be.