pinnacle_api::layout::generators

Struct Cycle

pub struct Cycle<T> {
    pub layouts: Vec<T>,
    /* private fields */
}

A LayoutGenerator that keeps track of layouts per tag and provides methods to cycle between them.

Fields
layouts: Vec<T>

The layouts this generator will cycle between.

Implementations

impl<T> Cycle<T>

pub fn new(layouts: impl IntoIterator<Item = T>) -> Self

Creates a new Cycle from the given LayoutGenerators.

Examples

let cycler = Cycle::new([
    Box::<MasterStack>::default() as Box<dyn LayoutGenerator + Send>,
    Box::<Dwindle>::default() as _,
    Box::<Corner>::default() as _,
]);

pub fn cycle_layout_forward(&mut self, tag: &TagHandle)

Cycles the layout forward on the given tag.

pub fn cycle_layout_backward(&mut self, tag: &TagHandle)

Cycles the layout backward on the given tag.

pub fn current_layout(&self, tag: &TagHandle) -> Option<&T>

Retrieves the current layout.

Returns None if no layouts were given.

pub fn set_current_tag(&mut self, tag: TagHandle)

Sets the current tag to choose a layout for.

pub fn current_tree_id(&self) -> u32

Gets a (most-likely) unique identifier for the current layout tree. This is guaranteed to be greater than zero.