Trait pinnacle_api::layout::LayoutGenerator

pub trait LayoutGenerator {
    // Required method
    fn layout(&self, args: &LayoutArgs) -> Vec<Geometry>;
}

Types that can generate layouts by computing a vector of geometries.

Required Methods

fn layout(&self, args: &LayoutArgs) -> Vec<Geometry>

Generate a vector of geometries using the given LayoutArgs.