pinnacle_api
Module output

Output management.

An output is Pinnacle’s terminology for a monitor.

This module provides Output, which allows you to get OutputHandles for different connected monitors and set them up.

Structs

Mode

A possible output pixel dimension and refresh rate configuration.

Modeline

A custom modeline.

Output

A struct that allows you to get handles to connected outputs and set them up.

OutputHandle

A handle to an output.

OutputProperties

The properties of an output.

OutputSetup

An output setup for use in Output::setup.

ParseModelineError

Error for the FromStr implementation for Modeline.

UpdateLocsOn

Flags for when Output::setup_locs should relayout outputs.

Enums

Alignment

The alignment to use for OutputHandle::set_loc_adj_to.

OutputId

An identifier for an output.

OutputLoc

A location for an output.

Transform

An output transform.