pinnacle_api::output
Struct Output

pub struct Output;

A struct that allows you to get handles to connected outputs and set them up.

See OutputHandle for more information.

Implementations

impl Output

pub fn get_all(&self) -> Vec<OutputHandle>

Get handles to all connected outputs.

Examples

let outputs = output.get_all();

pub async fn get_all_async(&self) -> Vec<OutputHandle>

The async version of Output::get_all.

pub fn get_all_enabled(&self) -> Vec<OutputHandle>

Get handles to all outputs that are connected and enabled.

Examples

let enabled = output.get_all_enabled();

pub async fn get_all_enabled_async(&self) -> Vec<OutputHandle>

The async version of Output::get_all_enabled.

pub fn get_by_name(&self, name: impl Into<String>) -> Option<OutputHandle>

Get a handle to the output with the given name.

By “name”, we mean the name of the connector the output is connected to.

Examples

let op = output.get_by_name("eDP-1")?;
let op2 = output.get_by_name("HDMI-2")?;

pub async fn get_by_name_async( &self, name: impl Into<String>, ) -> Option<OutputHandle>

The async version of Output::get_by_name.

pub fn get_focused(&self) -> Option<OutputHandle>

Get a handle to the focused output.

This is currently implemented as the one that has had the most recent pointer movement.

Examples

let op = output.get_focused()?;

pub async fn get_focused_async(&self) -> Option<OutputHandle>

The async version of Output::get_focused.

pub fn connect_for_all( &self, for_all: impl FnMut(&OutputHandle) + Send + 'static, )

Connect a closure to be run on all current and future outputs.

When called, connect_for_all will do two things:

Immediately run for_all with all currently connected outputs.

Create a future that will call for_all with any newly connected outputs.

Note that for_all will not run with outputs that have been unplugged and replugged. This is to prevent duplicate setup. Instead, the compositor keeps track of any tags and state the output had when unplugged and restores them on replug.

Examples

// Add tags 1-3 to all outputs and set tag "1" to active
output.connect_for_all(|op| {
    let tags = tag.add(&op, ["1", "2", "3"]);
    tags.first()?.set_active(true);
});

pub fn connect_signal(&self, signal: OutputSignal) -> SignalHandle

Connect to an output signal.

The compositor will fire off signals that your config can listen for and act upon. You can pass in an OutputSignal along with a callback and it will get run with the necessary arguments every time a signal of that type is received.

pub fn setup(&self, setups: impl IntoIterator<Item = OutputSetup>)

Declaratively setup outputs.

This method allows you to specify OutputSetups that will be applied to outputs already connected and that will be connected in the future. It handles the setting of modes, scales, tags, and more.

Setups will be applied top to bottom.

See OutputSetup for more information.

Examples

use pinnacle_api::output::OutputSetup;
use pinnacle_api::output::OutputId;

output.setup([
    // Give all outputs tags 1 through 5
    OutputSetup::new_with_matcher(|_| true).with_tags(["1", "2", "3", "4", "5"]),
    // Give outputs with a preferred mode of 4K a scale of 2.0
    OutputSetup::new_with_matcher(|op| op.preferred_mode()?.pixel_width == 2160)
        .with_scale(2.0),
    // Additionally give eDP-1 tags 6 and 7
    OutputSetup::new(OutputId::name("eDP-1")).with_tags(["6", "7"]),
]);

pub fn setup_locs( &self, update_locs_on: UpdateLocsOn, setup: impl IntoIterator<Item = (OutputId, OutputLoc)>, )

Specify locations for outputs and when they should be laid out.

This method allows you to specify locations for outputs, either as a specific point or relative to another output.

This will relayout outputs according to the given UpdateLocsOn flags.

Layouts not specified in setup or that have cyclic relative-to outputs will be laid out in a line to the right of the rightmost output.

Examples

use pinnacle_api::output::UpdateLocsOn;
use pinnacle_api::output::OutputLoc;
use pinnacle_api::output::OutputId;

output.setup_locs(
    // Relayout all outputs when outputs are connected, disconnected, and resized
    UpdateLocsOn::all(),
    [
        // Anchor eDP-1 to (0, 0) so other outputs can be placed relative to it
        (OutputId::name("eDP-1"), OutputLoc::Point(0, 0)),
        // Place HDMI-A-1 below it centered
        (
            OutputId::name("HDMI-A-1"),
            OutputLoc::RelativeTo(OutputId::name("eDP-1"), Alignment::BottomAlignCenter),
        ),
        // Place HDMI-A-2 below HDMI-A-1.
        (
            OutputId::name("HDMI-A-2"),
            OutputLoc::RelativeTo(OutputId::name("HDMI-A-1"), Alignment::BottomAlignCenter),
        ),
        // Additionally, if HDMI-A-1 isn't connected, place it below eDP-1 instead.
        (
            OutputId::name("HDMI-A-2"),
            OutputLoc::RelativeTo(OutputId::name("eDP-1"), Alignment::BottomAlignCenter),
        ),
    ]
);