pinnacle_api

Module process

Process management.

This module provides ways to spawn processes and handle their output.

Structs

Child

The result of spawning a Command.

Command

A process builder that allows you to spawn programs.

ExitInfo

Information from an exited process.

Functions

set_env

Adds an environment variable that all newly spawned Commands will inherit.