pinnacle_api

Module tag

Tag management.

This module allows you to interact with Pinnacle’s tag system.

The Tag System

Many Wayland compositors use workspaces for window management. Each window is assigned to a workspace and will only show up if that workspace is being viewed. This is a find way to manage windows, but it’s not that powerful.

Instead, Pinnacle works with a tag system similar to window managers like dwm and, the window manager Pinnacle takes inspiration from, awesome.

In a tag system, there are no workspaces. Instead, each window can be tagged with zero or more tags, and zero or more tags can be displayed on a monitor at once. This allows you to, for example, bring in your browsers on the same screen as your IDE by toggling the “Browser” tag.

Workspaces can be emulated by only displaying one tag at a time. Combining this feature with the ability to tag windows with multiple tags allows you to have one window show up on multiple different “workspaces”. As you can see, this system is much more powerful than workspaces alone.

Structs

TagHandle

A handle to a tag.

Functions

add

Adds tags to the specified output.

connect_signal

Connects to a TagSignal.

get

Gets a handle to the first tag with the given name on the focused output.

get_all

Gets handles to all tags across all outputs.

get_all_async

Async impl for get_all_async.

get_async

Async impl for get.

get_on_output

Gets a handle to the first tag with the given name on output.

get_on_output_async

Async impl for get_on_output.

remove

Removes the given tags from their outputs.