pinnacle_api::tag

Function get_on_output

pub fn get_on_output(
    name: impl ToString,
    output: &OutputHandle,
) -> Option<TagHandle>

Gets a handle to the first tag with the given name on output.

For a simpler way to get a tag on the focused output, see get.

Examples

let output = output::get_by_name("eDP-1")?;
let tag = tag::get_on_output("2", &output)?;