pinnacle_api::util
Macro batch_boxed

macro_rules! batch_boxed {
    [ $first:expr, $($request:expr),* ] => { ... };
}

A convenience macro to batch API calls in different concrete futures.

The batch function only accepts a collection of the same concrete future e.g. from a single async function or method.

To support different futures (that still return the same value), this macro will place provided futures in a Pin<Box<_>> to erase their type and pass them along to batch.

Examples

use pinnacle_api::util::batch_boxed;

let mut windows = window.get_all();
let first = windows.next()?;
let last = windows.last()?;

let classes: Vec<String> = batch_boxed![
    async {
        let class = first.class_async().await;
        class.unwrap_or("no class".to_string())
    },
    async {
        let mut class = last.class_async().await.unwrap_or("alalala");
        class += "hello";
        class
    },
];