Module pinnacle_api::window

Window management.

This module provides Window, which allows you to get WindowHandles and move and resize windows using the mouse.

WindowHandles allow you to do things like resize and move windows, toggle them between floating and tiled, close them, and more.

This module also allows you to set window rules; see the rules module for more information.

Modules

rules

Types for window rules.

Structs

Window

A struct containing methods that get WindowHandles and move windows with the mouse.

WindowHandle

A handle to a window.

WindowProperties

Properties of a window.

Enums

FullscreenOrMaximized

Whether a window is fullscreen, maximized, or neither.

WindowState

A window’s current display state.