pinnacle_api::window
Function for_each_window

pub fn for_each_window(for_all: impl FnMut(WindowHandle) + Send + 'static)

Runs a closure for every new window.

This function is how you implement window rules. Instead of a declarative window rule system with match conditions, you can use standard if statements and apply rules using the same methods that are used everywhere else in this API.

Note: this function is special in that if it is called, Pinnacle will wait for the provided closure to finish running before it sends windows an initial configure event. Do not block here. At best, short blocks will increase the time it takes for a window to open. At worst, a complete deadlock will prevent windows from opening at all.

Examples

use pinnacle_api::window;
use pinnacle_api::window::DecorationMode;
use pinnacle_api::tag;

window::for_each_window(|window| {
    // Make Alacritty always open on the "Terminal" tag
    if window.app_id() == "Alacritty" {
        window.set_tag(&tag::get("Terminal").unwrap(), true);
    }

    // Make all windows use client-side decorations
    window.set_decoration_mode(DecorationMode::ClientSide);
});