pinnacle_api::window

Struct WindowHandle

pub struct WindowHandle { /* private fields */ }

A handle to a window.

This allows you to manipulate the window and get its properties.

Implementations

impl WindowHandle

pub fn close(&self)

Sends a close request to this window.

If the window is unresponsive, it may not close.

pub fn set_geometry( &self, x: impl Into<Option<i32>>, y: impl Into<Option<i32>>, w: impl Into<Option<u32>>, h: impl Into<Option<u32>>, )

Sets this window’s location and/or size.

Only affects the floating geometry of windows. Tiled geometries are calculated using the current layout.

pub fn resize_tile(&self, left: i32, right: i32, top: i32, bottom: i32)

If this window is tiled, resizes its tile by shifting the left, right, top, and bottom edges by the provided pixel amounts.

Positive amounts shift edges right/down, while negative amounts shift edges left/up.

If this resizes the tile in a direction that it can no longer resize towards (e.g. it’s at the edge of the screen), it will resize in the opposite direction.

Examples

// Grow the focused tiled window 10 pixels leftward
window::get_focused()?.resize_tile(-10, 0, 0, 0);

// Shrink the focused tiled window 10 pixels inward from the right
window::get_focused()?.resize_tile(0, -10, 0, 0);

// Grow the focused tiled window 20 pixels centered vertically
window::get_focused()?.resize_tile(0, 0, -10, 10);

pub fn set_fullscreen(&self, set: bool)

Sets this window to fullscreen or not.

pub fn toggle_fullscreen(&self)

Toggles this window between fullscreen and not.

pub fn set_maximized(&self, set: bool)

Sets this window to maximized or not.

pub fn toggle_maximized(&self)

Toggles this window between maximized and not.

pub fn set_floating(&self, set: bool)

Sets this window to floating or not.

Floating windows will not be tiled and can be moved around and resized freely.

pub fn toggle_floating(&self)

Toggles this window to and from floating.

Floating windows will not be tiled and can be moved around and resized freely.

pub fn set_focused(&self, set: bool)

Focuses or unfocuses this window.

pub fn toggle_focused(&self)

Toggles this window between focused and unfocused.

pub fn set_decoration_mode(&self, mode: DecorationMode)

Sets this window’s decoration mode.

pub fn move_to_output(&self, output: &OutputHandle)

Moves this window to the specified output.

This will set the window tags to the output tags, and update the window position.

Example

// Move the focused window to output DP-2
window::get_focused()?.move_to_output(&output::get_by_name("DP-2")?);

pub fn move_to_tag(&self, tag: &TagHandle)

Moves this window to the given tag.

This will remove all tags from this window then tag it with tag, essentially moving the window to that tag.

Examples

// Move the focused window to tag "Code" on the focused output
window::get_focused()?.move_to_tag(&tag::get("Code")?);

pub fn set_tag(&self, tag: &TagHandle, set: bool)

Sets or unsets a tag on this window.

Examples

let focused = window::get_focused()?;
let tag = tag::get("Potato")?;

focused.set_tag(&tag, true); // `focused` now has tag "Potato"
focused.set_tag(&tag, false); // `focused` no longer has tag "Potato"

pub fn toggle_tag(&self, tag: &TagHandle)

Toggles a tag on this window.

Examples

let focused = window::get_focused()?;
let tag = tag::get("Potato")?;

focused.toggle_tag(&tag); // `focused` now has tag "Potato"
focused.toggle_tag(&tag); // `focused` no longer has tag "Potato"

pub fn set_tags<T: Borrow<TagHandle>>(&self, tags: impl IntoIterator<Item = T>)

Sets the exact provided tags on this window.

Passing in an empty collection will not change the window’s tags.

For ergonomics, this accepts iterators of both TagHandle and &TagHandle.

Examples

let focused = window::get_focused()?;
let tag1 = tag::get("1")?;
let tag3 = tag::get("3")?;

// Set `focused`'s tags to "1" and "3", removing all others
focused.set_tags([tag1, tag3]);

pub fn set_vrr_demand(&self, vrr_demand: impl Into<Option<VrrDemand>>)

Sets this window’s VrrDemand.

When set to None, this window has no vrr demand.

This works in conjunction with an output with Vrr::OnDemand.

pub fn raise(&self)

Raises this window to the front.

pub fn lower(&self)

Lowers this window to the back.

pub fn loc(&self) -> Option<Point>

Gets this window’s current location in the global space.

pub async fn loc_async(&self) -> Option<Point>

Async impl for Self::loc.

pub fn size(&self) -> Option<Size>

Gets this window’s current size.

pub async fn size_async(&self) -> Option<Size>

Async impl for Self::size.

pub fn app_id(&self) -> String

Gets this window’s app id (class if it’s an xwayland window).

If it doesn’t have one, this returns an empty string.

pub async fn app_id_async(&self) -> String

Async impl for Self::app_id.

pub fn title(&self) -> String

Gets this window’s title.

If it doesn’t have one, this returns an empty string.

pub async fn title_async(&self) -> String

Async impl for Self::title.

pub fn output(&self) -> Option<OutputHandle>

Gets this window’s output.

This is currently implemented as the output of the first tag that this window has.

Returns None if this window doesn’t exist or if it has no tags.

pub async fn output_async(&self) -> Option<OutputHandle>

Async impl for Self::output.

pub fn focused(&self) -> bool

Gets whether or not this window has keyboard focus.

pub async fn focused_async(&self) -> bool

Async impl for Self::focused.

pub fn layout_mode(&self) -> LayoutMode

Gets this window’s current LayoutMode.

pub async fn layout_mode_async(&self) -> LayoutMode

Async impl for Self::layout_mode.

pub fn floating(&self) -> bool

Gets whether or not this window is floating.

pub async fn floating_async(&self) -> bool

Async impl for Self::floating.

pub fn tiled(&self) -> bool

Gets whether or not this window is tiled.

pub async fn tiled_async(&self) -> bool

Async impl for Self::tiled.

pub fn fullscreen(&self) -> bool

Gets whether or not this window is fullscreen.

pub async fn fullscreen_async(&self) -> bool

Async impl for Self::fullscreen.

pub fn maximized(&self) -> bool

Gets whether or not this window is maximized.

pub async fn maximized_async(&self) -> bool

Async impl for Self::maximized.

pub fn tags(&self) -> impl Iterator<Item = TagHandle> + use<>

Gets handles to all tags on this window.

pub async fn tags_async(&self) -> impl Iterator<Item = TagHandle> + use<>

Async impl for Self::tags.

pub fn is_on_active_tag(&self) -> bool

Gets whether or not this window has an active tag.

pub async fn is_on_active_tag_async(&self) -> bool

Async impl for Self::is_on_active_tag.

pub fn in_direction( &self, direction: Direction, ) -> impl Iterator<Item = WindowHandle> + use<>

Gets all windows in the provided direction, sorted closest to farthest.

pub async fn in_direction_async( &self, direction: Direction, ) -> impl Iterator<Item = WindowHandle> + use<>

Async impl for Self::in_direction.

pub fn foreign_toplevel_list_identifier(&self) -> Option<String>

Gets this window’s ext-foreign-toplevel-list handle identifier.

pub async fn foreign_toplevel_list_identifier_async(&self) -> Option<String>

Async impl for Self::foreign_toplevel_list_identifier.

pub fn swap(&self, target: &WindowHandle)

Swap position with another window.

pub async fn swap_async(&self, target: &WindowHandle)

Async impl for Self::swap.

pub fn id(&self) -> u32

Gets this window’s raw compositor id.

pub fn from_id(id: u32) -> Self

Creates a window handle from an ID.

Note: This is mostly for testing and if you want to serialize and deserialize window handles.