pinnacle_api

Module debug

Debugging utilities.

WARNING: This module is not governed by the API stability guarantees.

Functions

set_cursor_plane_scanout

Enables or disables cursor plane scanout.

set_damage_visualization

Sets damage visualization.

set_opaque_region_visualization

Sets opaque region visualization.

set_process_piping

Enables or disables process spawning setting up pipes to expose fds to the config.

toggle_cursor_plane_scanout

Toggles cursor plane scanout.

toggle_damage_visualization

Toggles damage visualization.

toggle_opaque_region_visualization

Toggles opaque region visualization.

toggle_process_piping

Toggles process spawning setting up pipes to expose fds to the config.