pinnacle_api

Module input

Input management.

This module provides ways to manage bindings, input devices, and other input settings.

Modules

libinput

Libinput device configuration.

Structs

BindInfo

Bind information.

BindLayer

A bind layer, also known as a bind mode.

Keybind

A keybind.

Keysym

The keyboard symbol, often corresponding to a character.

Mod

A keyboard modifier for use in binds.

Mousebind

A mousebind.

XkbConfig

A struct that lets you define xkeyboard config options.

Enums

BindInfoKind

The kind of a bind (hey that rhymes).

MouseButton

A mouse button.

Traits

Bind

Functionality common to all bind types.

ToKeysym

A trait that designates anything that can be converted into a Keysym.

Functions

bind_infos

Gets all bind information.

connect_signal

Connects to an InputSignal.

cycle_xkb_layout_backward

Cycles the current XKB layout backward.

cycle_xkb_layout_forward

Cycles the current XKB layout forward.

keybind

Creates a keybind on the DEFAULT bind layer.

mousebind

Creates a mousebind on the DEFAULT bind layer.

set_repeat_rate

Sets the keyboard’s repeat rate.

set_xcursor_size

Sets the xcursor size.

set_xcursor_theme

Sets the xcursor theme.

set_xkb_config

Sets the xkeyboard config.

set_xkb_keymap

Sets the XKB keymap.

switch_xkb_layout

Switches the current XKB layout to the one at the provided index.