pinnacle_api::input

Module libinput

Libinput device configuration.

Structs

Capability

A device’s libinput capabilities.

DeviceHandle

A libinput device.

Enums

AccelProfile

A pointer acceleration profile.

ClickMethod

The click method defines when to generate software-emulated buttons, usually on a device that does not have a specific physical button available.

DeviceType

A device’s type.

ScrollMethod

The scroll method of a device selects when to generate scroll axis events instead of pointer motion events.

SendEventsMode

A libinput send events mode.

TapButtonMap

Map 1/2/3 finger taps to buttons.

Functions

for_each_device

Runs a closure for all current and future input devices.

get_devices

Gets handles to all connected input devices.