pinnacle_api
Module output

Output management.

An output is the Wayland term for a monitor. It presents windows, your cursor, and other UI elements.

Outputs are uniquely identified by their name, a.k.a. the name of the connector they’re plugged in to.

Structs

Mode

A possible output pixel dimension and refresh rate configuration.

Modeline

A custom modeline.

OutputHandle

A handle to an output.

ParseModelineError

Error for the FromStr implementation for Modeline.

Enums

Alignment

The alignment to use for OutputHandle::set_loc_adj_to.

Transform

An output transform.

Functions

connect_signal

Connects to an OutputSignal.

for_each_output

Runs a closure on all current and future outputs.

get_all

Gets handles to all currently plugged-in outputs.

get_all_async

Async impl for get_all.

get_all_enabled

Gets handles to all currently plugged-in and enabled outputs.

get_all_enabled_async

Async impl for get_all_enabled.

get_by_name

Gets a handle to the output with the given name.

get_by_name_async

Async impl for get_by_name.

get_focused

Gets a handle to the currently focused output.

get_focused_async

Async impl for get_focused.