Function pinnacle_api::util::batch

pub fn batch<T>(
    requests: impl IntoIterator<Item = impl Future<Output = T>>
) -> Vec<T>

Batch a set of requests that will be sent ot the compositor all at once.

Rationale

Normally, all API calls are blocking. For example, calling Window::get_all then calling WindowHandle.props on each returned window handle will block after each props call waiting for the compositor to respond:

// This will block after each call to `window.props()`, meaning this all happens synchronously.
// If the compositor is running slowly for whatever reason, this will take a long time to complete.
let props = window.get_all()
    .map(|window| window.props())
    .collect::<Vec<_>>();

In order to mitigate this issue, you can batch up a set of API calls using this function. This will send all requests to the compositor at once without blocking, then wait for the compositor to respond.

You’ll see that this function takes in an IntoIterator of Futures. As such, all API calls that return something have an async variant named *_async that returns a future. You must pass these futures into the batch function instead of their non-async counterparts.

The batch_boxed macro

The util module also provides the batch_boxed macro.

The batch function only accepts one concrete type of future, meaning that you can only batch a collection of futures from one specific function or method.

As a convenience, batch_boxed accepts one or more different futures that return the same type. It will place provided futures in a Pin<Box<_>> to erase the types and pass them along to batch.

Examples

use pinnacle_api::util::batch;
use pinnacle_api::window::WindowProperties;

let props: Vec<WindowProperties> = batch(window.get_all().map(|window| window.props_async()));
                                                        // Don't forget the `async` ^^^^^