pinnacle_api

Module window

Window management.

This module provides ways to get WindowHandles and move and resize windows using the mouse.

WindowHandles allow you to do things like resize and move windows, toggle them between floating and tiled, close them, and more.

Structs

VrrDemand

A demand for variable refresh rate on an output.

WindowHandle

A handle to a window.

Enums

DecorationMode

A mode for window decorations (titlebar, shadows, etc).

LayoutMode

A window’s current layout mode.

Functions

add_window_rule

Adds a window rule.

begin_move

Begins an interactive window move.

begin_resize

Begins an interactive window resize.

connect_signal

Connects to a WindowSignal.

get_all

Gets handles to all windows.

get_all_async

Async impl for get_all.

get_focused

Gets a handle to the window with the current keyboard focus.

get_focused_async

Async impl for get_focused.