Module pinnacle_api::window::rules

Types for window rules.

A window rule is a way to set the properties of a window on open.

They are comprised of two parts: the condition and the actual rule.

WindowRuleConditions

WindowRuleConditions are conditions that the window needs to open with in order to apply a rule. For example, you may want to set a window to maximized if it has the class “steam”, or you might want to open all Firefox instances on tag “3”.

To do this, you must build a WindowRuleCondition to tell the compositor when to apply any rules.

Building WindowRuleConditions

A condition is created through WindowRuleCondition::new:

let cond = WindowRuleCondition::new();

In order to understand conditions, you must understand the concept of “any” and “all”.

“Any”

“Any” conditions only need one of their constituent items to be true for the whole condition to evaluate to true. Think of it as one big if a || b || c || d || ... {} block.

“All”

“All” conditions need all of their constituent items to be true for the condition to evaluate to true. This is like a big if a && b && c && d && ... {} block.

Note that any items in a top level WindowRuleCondition fall under “all”, so all those items must be true.

With that out of the way, we can get started building conditions.

WindowRuleCondition::classes

With WindowRuleCondition::classes, you can specify what classes a window needs to have for a rule to apply.

The following will apply to windows with the class “firefox”:

let cond = WindowRuleCondition::new().classes(["firefox"]);

Note that you pass in some impl IntoIterator<Item = impl Into<String>>. This means you can pass in more than one class here:

let failing_cond = WindowRuleCondition::new().classes(["firefox", "steam"]);

HOWEVER: this will not work. Recall that top level conditions are implicitly “all”. This means the above would require windows to have both classes, which is impossible. Thus, the condition above will never be true.

WindowRuleCondition::titles

Like classes, you can use titles to specify that the window needs to open with a specific title for the condition to apply.

let cond = WindowRuleCondition::new().titles(["Steam"]);

Like classes, passing in multiple titles at the top level will cause the condition to always fail.

WindowRuleCondition::tags

You can specify that the window needs to open on the given tags in order to apply a rule.

let cond = WindowRuleCondition::new().tags([&tag.get("3", output.get_by_name("HDMI-1")?)?]);

Here, if you have tag “3” active on “HDMI-1” and spawn a window on that output, this condition will apply.

Unlike classes and titles, you can specify multiple tags at the top level:

let op = output.get_by_name("HDMI-1")?;
let tag1 = tag.get("1", &op)?;
let tag2 = tag.get("2", &op)?;

let cond = WindowRuleCondition::new().tags([&tag1, &tag2]);

Now, you must have both tags “1” and “2” active and spawn a window for the condition to apply.

WindowRuleCondition::any

Now we can get to ways to compose more complex conditions.

WindowRuleCondition::any takes in conditions and will evaluate to true if anything in those conditions are true.

let cond = WindowRuleCondition::new()
    .any([
        WindowRuleCondition::new().classes(["Alacritty"]),
        WindowRuleCondition::new().tags([&tag.get("2", None)?]),
    ]);

This condition will apply if the window is either “Alacritty” or opens on tag “2”.

WindowRuleCondition::all

With WindowRuleCondition::all, all specified conditions must be true for the condition to be true.

let cond = WindowRuleCondition::new()
    .all([
        WindowRuleCondition::new().classes(["Alacritty"]),
        WindowRuleCondition::new().tags([&tag.get("2", None)?]),
    ]);

This condition applies if the window has the class “Alacritty” and opens on tag “2”.

You can write the above a bit shorter, as top level conditions are already “all”:

let cond = WindowRuleCondition::new()
    .classes(["Alacritty"])
    .tags([&tag.get("2", None)?]);

Complex condition composition

You can arbitrarily nest any and all to achieve desired logic.

let op = output.get_by_name("HDMI-1")?;
let tag1 = tag.get("1", &op)?;
let tag2 = tag.get("2", &op)?;

let complex_cond = WindowRuleCondition::new()
    .any([
        WindowRuleCondition::new().all([
            WindowRuleCondition::new()
                .classes("Alacritty")
                .tags([&tag1, &tag2])
        ]),
        WindowRuleCondition::new().all([
            WindowRuleCondition::new().any([
                WindowRuleCondition::new().titles(["nvim", "emacs", "nano"]),
            ]),
            WindowRuleCondition::new().any([
                WindowRuleCondition::new().tags([&tag1, &tag2]),
            ]),
        ])
    ])

The above is true if either of the following are true:

The window has class “Alacritty” and opens on both tags “1” and “2”, or

The window’s class is either “nvim”, “emacs”, or “nano” and it opens on either tag “1” or “2”.

WindowRules

WindowRuleConditions are half of a window rule. The other half is the WindowRule itself.

A WindowRule is what will apply to a window if a condition is true.

Building WindowRules

Create a new window rule with WindowRule::new:

let rule = WindowRule::new();

There are several rules you can set currently.

WindowRule::output

This will cause the window to open on the specified output.

WindowRule::tags

This will cause the window to open with the given tags.

WindowRule::floating

This will cause the window to open either floating or tiled.

WindowRule::fullscreen_or_maximized

This will cause the window to open either fullscreen, maximized, or neither.

WindowRule::x

This will cause the window to open at the given x-coordinate.

Note: this only applies to floating windows; tiled windows’ geometry will be overridden by layouting.

WindowRule::y

This will cause the window to open at the given y-coordinate.

Note: this only applies to floating windows; tiled windows’ geometry will be overridden by layouting.

WindowRule::width

This will cause the window to open with the given width in pixels.

Note: this only applies to floating windows; tiled windows’ geometry will be overridden by layouting.

WindowRule::height

This will cause the window to open with the given height in pixels.

Note: this only applies to floating windows; tiled windows’ geometry will be overridden by layouting.

Structs

WindowRule

A window rule.

WindowRuleCondition

A condition for a WindowRule to apply to a window.